Refactor dev

+++ tests/test_combiner.py
@@ -6,7 +6,7 @@ import ruamel.yaml
 import tensorflow as tf
 
 from model.combiner import Combiner
-from preprocessing.preprocessor import Preprocessor
+from utils.preprocessing.preprocessor import Preprocessor
 
 
 class TestCombiner(unittest.TestCase):
@@ -19,7 +19,7 @@ class TestCombiner(unittest.TestCase):
         yaml = ruamel.yaml.YAML()
         with open(config_path, 'r') as f:
             self.config = yaml.load(f)
-
+    
     def test_training(self):
         test_mels = [np.random.random((100 + i * 5, 80)) for i in range(10)]
         combiner = Combiner(self.config)
@@ -30,10 +30,10 @@ class TestCombiner(unittest.TestCase):
         train_samples = [preprocessor.encode('repeated text', mel) for mel in test_mels]
         train_set_gen = lambda: (item for item in train_samples)
         train_dataset = tf.data.Dataset.from_generator(train_set_gen,
-                                                       output_types=(tf.float32, tf.int64, tf.int64))
+                                                       output_types=(tf.float32, tf.int32, tf.int32))
         train_dataset = train_dataset.shuffle(10).padded_batch(
             self.config['batch_size'], padded_shapes=([-1, 80], [-1], [-1]), drop_remainder=True)
-
+        
         outputs = []
         for epoch in range(self.config['epochs']):
             for (batch, (mel, text, stop)) in enumerate(train_dataset):
@@ -43,12 +43,12 @@ class TestCombiner(unittest.TestCase):
                                              pre_dropout=0.5,
                                              mask_prob=self.config['mask_prob'])
                 outputs.append(output)
-
+        
         self.assertAlmostEqual(2.2329392433166504, float(outputs[-1]['text_mel']['loss']), places=6)
-        self.assertAlmostEqual(2.265047550201416 , float(outputs[-1]['mel_mel']['loss']), places=6)
+        self.assertAlmostEqual(2.265047550201416, float(outputs[-1]['mel_mel']['loss']), places=6)
         self.assertAlmostEqual(0.0014063421403989196, float(outputs[-1]['mel_text']['loss']), places=6)
         self.assertAlmostEqual(0.0011274153366684914, float(outputs[-1]['text_text']['loss']), places=6)
-
+        
         mel_input, text_input = train_samples[0][0], train_samples[0][1]
         pred_mel_text = combiner.mel_text.predict(mel_input, max_length=10)
         pred_text_text = combiner.text_text.predict(text_input, max_length=10)
@@ -58,4 +58,3 @@ class TestCombiner(unittest.TestCase):
         self.assertAlmostEqual(-19.69157600402832, float(tf.reduce_sum(pred_text_text['logits'])))
         self.assertAlmostEqual(-786.9090576171875, float(tf.reduce_sum(pred_text_mel['mel'])))
         self.assertAlmostEqual(-787.0718383789062, float(tf.reduce_sum(pred_mel_mel['mel'])))
-
